Fix CPP compensation amount calc logic during payment amount updating
Previously if the "increaseCashback()" function of the CashbackDistributor contract failed the compensation amount of a payment increased nevertheless.
Now if the "increaseCashback()" function fails the compensation amount of a payment remains the same.

Note. Line 273 has been moved from its original place just to keep the symmetry of the two parts of the code in the related if-else statement.